refactor: Move ImmutablesStyle to core module for easier reuse
