+++ core/src/main/java/io/syndesis/core/immutable/ImmutablesStyle.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.syndesis.model;
+package io.syndesis.core.immutable;
 
 import org.immutables.value.Value;
