+++ model/src/main/java/io/syndesis/model/connection/package-info.java
@@ -16,4 +16,4 @@
 @ImmutablesStyle
 package io.syndesis.model.connection;
 
-import io.syndesis.model.ImmutablesStyle;
+import io.syndesis.core.immutable.ImmutablesStyle;
