+++ model/src/main/java/io/syndesis/model/integration/package-info.java
@@ -16,4 +16,4 @@
 @ImmutablesStyle
 package io.syndesis.model.integration;
 
-import io.syndesis.model.ImmutablesStyle;
+import io.syndesis.core.immutable.ImmutablesStyle;
