+++ project-generator/src/main/java/io/syndesis/project/converter/package-info.java
@@ -16,4 +16,4 @@
 @ImmutablesStyle
 package io.syndesis.project.converter;
 
-import io.syndesis.model.ImmutablesStyle;
+import io.syndesis.core.immutable.ImmutablesStyle;
